Updated the way new clients join rooms
Clients send the room ID of the room they want to join.

From the client-side, a client can either send the room ID of an existing room, or create a random room ID and send that, in which case a new room with that ID is created by the server.