Added functions for executing Python and JavaScript code
Now the client will be able to send an "executeCode" event via the socket connection, which will contain the code and programming  language.

The server can then store the code into a file and then execute it, and send the output of the code, along with the errors, if any, back to the clients in the room, as a "codeOutput" event.

Currently, only Python 3 or JavaScript code may be executed on the server.